+++ src/main/java/com/mee/manage/po/DataTop.java
@@ -13,7 +13,7 @@ public class DataTop {
     @TableId(type = IdType.ID_WORKER)
     Long id;
 
-    Integer bizId;
+    Long bizId;
     
     String sku;
